FinalBrainWave/MODEL_GCP/requirements.txt
absl-py==1.4.0
apache-beam==2.46.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.7.3
fasteners==0.18
flatbuffers==23.3.3
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.16.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-dlp==3.12.1
google-cloud-language==1.3.2
google-cloud-pubsub==2.16.0
google-cloud-pubsublite==1.8.1
google-cloud-recommendations-ai==0.7.1
google-cloud-spanner==3.31.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.48.2
h5py==3.8.0
hdfs==2.7.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.6.0
jax==0.4.8
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opt-einsum==3.3.0
orjson==3.8.10
overrides==6.5.0
packaging==23.1
pandas==1.5.3
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==6.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
pymongo==3.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.1
tensorflow-serving-api==2.10.1
tensorflow-transform==1.13.0
termcolor==2.3.0
tfx-bsl==1.13.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.15
Werkzeug==2.3.0
wrapt==1.14.1
zipp==3.15.0
zstandard==0.21.0
